feat: attach blocks to delegation (#27)
This PR adds a method `Attach` to `Delegation` that allows additional
blocks to be attached. The primary usecase is to allow blocks linked
from capabilities or facts to be included in the delegation payload.

Note: this feature exists in JS Ucanto already and it actually validates
the blocks you attach are referenced from capabilities/facts. I will
open an issue to track feature parity here assuming this is approved.