feat: support invocations as effects (#30)
This PR allows completed effects to be included in a receipt. Similar to
[`Ran`](https://github.com/storacha/go-ucanto/blob/main/core/invocation/ran/ran.go),
an effect is now a union type - a link OR an invocation.

This allows a location commitment to be returned from a `blob/accept`
invocation. 🤔 Not sure I agree with using effects for this, but this is
how it is done currently.