feat: create ed25519 and RSA signer/verifier from raw public/private key bytes (#34)
This PR adds a `FromRaw([]byte)` function to ed25519 and RSA signers and
verifiers. It is a compliment to the `.Raw()` interface function that
`Signer` and `Verifier` has. It allows to round trip to/from raw bytes
(non-multiformats) and also create signers and verifiers from compatible
raw public/private key bytes generated from other ed25519/RSA libraries.

For example, it allows an ed25519 private key created by the libp2p
crypto library to be converted to a Ucanto signer.