fix: export includes only delegation blocks (#45)
Fixes `Blocks()` function of a `Delegation` to include only the root
block of the delegation, any (non-absent) proofs and any attached
blocks. This is [consistent with the JS
implementation](https://github.com/storacha/ucanto/blob/46612a897f3981dba272d19367fb75c6720a70cd/packages/core/src/delegation.js#L500).

Previously exporting delegation blocks would include ALL blocks in the
blockstore used to instantiate the delegation. This prevents us from
instantiating a delegation from a blockstore that contains multiple
unrelated delegations or other blocks not relevant to the delegation.

As a bonus, the constructor is altered to fetch and decode the root
block allowing it to fail with an error instead of failing later on
property access and possibly causing a npe panic.