feat: custom failures (#55)
This PR changes the server handler return types.

Previously a handler signature was:

```go
func[O any](Context, Capability[C], Invocation, InvocationContext) (O, fx.Effects, error)
```

It is now:

```go
func[O any, X failure.IPLDBuilderFailure](Context, Capability[C], Invocation, InvocationContext) (result.Result[O, X], fx.Effects, error)
```

i.e. Instead of returning a success value (`O`), it now returns a
`result.Result[O, X]`.

This allows the handler to return a custom error in a `result.Result`.
This is in the case where the error is _known_. Previously all errors
would have a `name` of `HandlerExecutionError` and the actual error
would be communicated in the `cause` property. This is awkward to unpack
on the client, is redundant (every error is currently a
`HandlerExecutionError`) and incompatible with the JS implementation.

The intention of `HandlerExecutionError` is a catch all for _unexpected
errors_, however it is currently being used to communicate _all_
failures. This is my fault for rushing the implementation.

The failure type `X` is constrained to `failure.IPLDBuilderFailure`,
which ensures known errors a) have a `name` property, b) are an `error`
(and have a message) and c) have a `ToIPLD()` method.

To summarise, handlers should communicate _known_ errors as an error
result, and unexpected errors in the `error` return value.

Note: there is an alternative to this
https://github.com/storacha/go-ucanto/pull/56