feat!: retrieval server and client (#48)
This PR implements the [RFC
here](https://github.com/storacha/RFC/pull/60). It exposes a server and
client implementation that allows UCAN authorized retrieval requests via
invocations (and receipts) passed in HTTP headers. This leaves the HTTP
response body available to be used for retrieved bytes.

A retrieval server is very similar to a normal Ucanto server, except it
requires invocations to be sent using the `headercar` transport codec.
The only other difference is that invocation handlers receive an extra
argument - the HTTP request info, and can return and additional value -
a HTTP response.

The retrieval client is also very similar to a Ucanto client, except
that it has the ability to send an invocation in multiple parts, if it
does not fit in HTTP headers. Essentially it'll send proofs one by one
until the server has all the proofs required to execute the invocation.
The server has an LRU cache allowing for this.

The PR also includes a transport codec that encodes agent messages into
HTTP headers.

🎬 Demo: https://youtu.be/11np-cGTe48?si=kw88R1DAlMSq-b1T

resolves #59

---------

Co-authored-by: Forrest <forrest@storacha.network>