feat: receipts in the server can be logged (#67)
Ref: https://github.com/storacha/piri/issues/174

When client code registers UCAN handlers in a go-ucanto server, there is
currently no way for that code to grasp the receipt the server generates
and sends back to the invoker. This PR adds a new `WithRequestLogger`
server option to register a callback the server will use to pass the
receipt back.

Additionally, a new `WithInvocation` method is added to the `Receipt`
type that allows producing a receipt with an attached invocation from
one that doesn't have it. Receipts issued by the retrieval server don't
include invocation blocks to save space in the headers, but it is useful
that logged receipts are full receipts in the sense that they also have
the invocation blocks in them.