fix: include nonce and notBefore when validating signatures (#77)
The nonce and nbf fields, when set, are included in the signature
payload when calculating the signature when a new UCAN is issued.
However, they are not included in the payload when verifying signatures.
This means that any UCAN with a nonce or a nbf field set will fail
signature validation.

We likely haven't seen this happening in the wild because we:
 - usually don't set nonces
- we do use the nbf field, but validating it happens before validating
the signature, so it fails earlier. I found this issue when working on
adding an option to ignore expiration and nbf fields on validation.

Note: the test I added failed before the fix.