feat: make time bounds validation configurable (#78)
Expiration and Not Valid Before are useful when validating an invocation
as part of deciding whether it should be executed. However, if we just
want to check the validity of a given delegation chain that happened in
the past, we don't care about time bounds.

These changes allow passing a custom time bounds validation function to
decide how expiration and not valid before are validated.